Address bad_numeric_casts for filetransfers
The filetransfer blockSize is now an unsigned integer, as 0 could be
used to denote an invalid block size as well (and indeed, already
indicated that better than -1 did).

All use of numeric_cast in filetransfer code has been fixed to deal with
the possibility of thrown exceptions.

Test-Information:

Unit tests pass on macOS and Debian

+++ Swiften/FileTransfer/IBBSendSession.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2016 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -40,7 +40,7 @@ IBBSendSession::~IBBSendSession() {
 
 void IBBSendSession::start() {
     IBBRequest::ref request = IBBRequest::create(
-            from, to, IBB::createIBBOpen(id, boost::numeric_cast<int>(blockSize)), router);
+            from, to, IBB::createIBBOpen(id, blockSize), router);
     request->onResponse.connect(boost::bind(&IBBSendSession::handleIBBResponse, this, _1, _2));
     active = true;
     request->send();
